requirements.txt
python-docx
openai
tqdm
tenacity 
ratelimit